Daemon support (TCP version)
faked goes into "detached" mode when it receives SIGHUP: it will terminate
as soon as all clients close their connections.  The fakeroot script kills
faked with SIGHUP instead of SIGTERM when it exits.  This is done in order
to provide a working (although isolated) fakeroot environment for daemon
processes that are launched from a fakeroot session.

git-archimport-id: fakeroot@packages.debian.org--fakeroot/fakeroot--main--0.0--patch-69