Wrap sysinfo(3) on Solaris in order to work around ISA issues
Solaris provides a way to select binary basing on the available ISA, which
breaks fakeroot running any non-trivial, as a mixture of 64-bit and 32-bit
binaries are being executed. Unlike Linux, LD_PRELOAD'ed library of wrong
architecture is not a warning, but a fatal error, and dynamic linker refuses to
start instead of ignoring the library.

In order to avoid this problem, fix the list of ISAs returned by sysinfo(3) to
the ones compatible with the fakeroot's own binary.