Fix packagemaker failures on OS X when running 10.6+
The packagemaker utility spawns a child to help with the package
creation but it explicitly sets its own insert library for that
child.  Previously this was handled by intercepting only execve,
but on newer systems that's not good enough.  Now intercept all
the exec/spawn calls to restore the insert library.