configure.ac: fix __xmknod{,at} pointer argument
Switch default to assume * and not the absence of *.

On glibc 2.33+, there is no definition for these functions in header
files, so the compile test doesn't work. But, we can default to using
the pointer (as is the case with newer glibc), and use the header file
on older platforms to fail the test and use no pointer.

Signed-off-by: Ilya Lipnitskiy <ilya.lipnitskiy@gmail.com>